requirements.txt
selenium
pytest
pytest-html
pyyaml
requests
